requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
PyJWT==2.9.0
PyMuPDF==1.25.5
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
